+++ src/addons/display_commander/ui/new_ui/debug/ngx_counters_tab.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+// Source Code <Display Commander>
+
+#include "../../../ui/imgui_wrapper_base.hpp"
+
+namespace ui::new_ui::debug {
+
+void DrawNGXCountersTab(display_commander::ui::IImGuiWrapper& imgui);
+
+}  // namespace ui::new_ui::debug